pyproject.toml
# pyproject.toml
[build-system]
requires = ["poetry-core>=1.2.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "fastapi-githubapp"
version = "0.2.5"
description = "FastAPI extension for rapid GitHub App development"
readme = "README.md"
authors = ["primetheus <865381+primetheus@users.noreply.github.com>"]
license = "MIT"
homepage = "https://github.com/primetheus/fastapi-githubapp"
repository = "https://github.com/primetheus/fastapi-githubapp"
keywords = [
  "fastapi",
  "github",
  "app",
  "oauth2",
  "oauth",
  "probot",
  "githubapp",
  "github-app"
]
classifiers = [
  "Programming Language :: Python :: 3",
  "Framework :: FastAPI",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent"
]

# point Poetry at our src/ layout
packages = [
  { include = "githubapp", from = "src" }
]

# Tell Poetry to include README and LICENSE in your sdist:
include = ["README.md", "LICENSE"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = ">=0.95.0"
uvicorn = { version = ">=0.22.0", extras = ["standard"] }
ghapi = ">=1.0.0"
pyjwt = { version = ">=2.8.0", extras = ["crypto"] }
httpx = ">=0.28.1"

[tool.poetry.group.dev.dependencies]
pytest = "*"
pytest-mock = "*"
isort = "*"
black = "*"
pre-commit = "*"
pylint = "*"
seed-isort-config = "*"
green = "*"
tox = "*"
tox-gh-actions = "*"
respx = ">=0.21.0"
python-semantic-release = "^10.0.2"

[tool.semantic_release]
branch = "main"
tag_format = "v{version}"
version_variable = "pyproject.toml:tool.poetry.version"
upload_to_github = true
upload_to_pypi = true

[tool.pytest.ini_options]
# run tests in the "tests" folder
testpaths = ["tests"]
# add your src/ tree so "import githubapp" works
pythonpath = ["src"]
# pick up only files named test_*.py
python_files = ["test_*.py"]
# shorten output, show extra summary info for skips/fails
addopts = "-q --disable-warnings --strict-markers"
